requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
pydantic==2.5.3
langgraph==0.0.26
pyyaml==6.0.1
typing-extensions==4.9.0